pyproject.toml
[project]
name = "nldate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dateutil>=2.9.0.post0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/nldate"]

[dependency-groups]
dev = [
    "mypy>=2.1.0",
    "pytest>=9.0.3",
    "ruff>=0.15.12",
    "types-python-dateutil>=2.9.0.20260508",
]